move node.py into folder and split into files (#82)
